backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
psycopg2-binary
langchain
langchain-openai
langchain-community
langchain-core
langgraph
openai
sqlalchemy
python-multipart